+++ seed/template/all/angularjs/compiler/babeljs/mocha/component/main.spec.js
@@ -1,0 +1,26 @@
+describe('Component: <%= component %>', () => {
+    beforeEach(module('<%= project %>'))
+
+    let component
+    let ctrl
+    let scope
+
+    beforeEach(inject(($rootScope, $compile) => {
+        scope = $rootScope.$new()
+
+        component = angular.element('<<%= component %>></<%= component %>>')
+        component = $compile(component)(scope)
+
+        ctrl = component.controller('<%= component %>')
+
+        scope.$apply()
+    }))
+
+    it('Text in Template', () => {
+        assert.equal(component.find('div').text(), 'Hello Component <%= component %>')
+    })
+
+    it('Property Ctrl', () => {
+        assert.equal(ctrl.hello, undefined)
+    })
+})