+++ seed/template/test/mocha/karma.conf.js
@@ -7,9 +7,9 @@ module.exports = function (config) {
         files: $.test.files,
         preprocessors: $.test.preprocessors,
 
-        singleRun: true,
+        singleRun: $.test.singleRun,
 
-        browsers: ['PhantomJS'],
+        browsers: $.test.browsers,
 
         frameworks: [
             'chai',
@@ -19,6 +19,7 @@ module.exports = function (config) {
         plugins: [
             'karma-chai',
             'karma-mocha',
+            'karma-chrome-launcher',
             'karma-phantomjs-launcher'
         ]
     }