+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 - [Spanish](#spanish)
 
 ## English
+### v.0.17.0
+    - Fix: npm run component
+    - Fix: watch in *.spec.js
+    - .core-config.json to core-config.js
+    - Add Tests in npm run component with AngularJS
+    - Add vendor in property test.files of core-config.js
+    - Allow use Chrome or PhantomJS in Karma
+    - Simplifies core-config.js
+
 ### v.0.16.0
     - Fix: npm run test 
     - Allow choice options to show in templates
@@ -186,6 +195,15 @@
     - Restart changelog in 2 languages
 
 ## Spanish
+### v.0.17.0
+    - Arregla: npm run component
+    - Arregla: watch en *.spec.js
+    - .core-config.json a core-config.js
+    - Añade Test en npm run component con AngularJS
+    - Añade vendor en propiedad test.files de core-config.js
+    - Permite usar Chrome o PhantomJS en Karma
+    - Simplifica core-config.js
+
 ### v.0.16.0
     - Arregla: npm run test 
     - Permite elegir opciones a mostrar en plantillas